src/lib.rs,src/interpreter.rs: Move interpreter to dedicated module
The interpreter is rather self-contained, assuming we can pass it the
program and the list of helpers to use. Let's move it to a dedicated
module, to lighten lib.rs.

No functional change.
